requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
aiohttp==3.8.5
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.2
pytest-asyncio==0.21.1
pydantic==2.3.0 